Fixing Windows build error

+++ src/runtime_src/xocl/api/plugin/xdp/profile_counters.cpp
@@ -138,7 +138,7 @@ namespace {
     auto& cumask = xrt_core::kernel_int::get_cumask(run) ;
     for (size_t bit = 0; bit < cumask.size(); ++bit)
       if (cumask.test(bit))
-	return bit;
+	return static_cast<unsigned int>(bit);
     return 0 ;
   }
